Correct crew validation rules
The validation rules for `user_id` were causing updating the crew role to fail as no `user_id` field is sent by the AJAX request.

This corrects the problem by only using the `exists` validation rule if the form is creating a new crew role. If a role is being edited, the `user_id` isn't changed so no validation is required.